Revert "Merge remote-tracking branch 'mirror/trunk' into backward-bmc-algorithm-local"
This reverts commit c291b38c64451af461b3e6aea6786d08182a6559.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/backward-bmc-algorithm@44014 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/ci/translators/TranslatorTest.java
@@ -152,12 +152,12 @@ public class TranslatorTest {
     PersistentMap<String, Interval> intervals = PathCopyingPersistentTreeMap.of();
     PersistentMap<String, Integer> referenceMap = PathCopyingPersistentTreeMap.of();
 
-    intervals = intervals.putAndCopy("var1", new Interval(Long.MIN_VALUE, 5L));
-    intervals = intervals.putAndCopy("var2", new Interval(-7L, Long.MAX_VALUE));
-    intervals = intervals.putAndCopy("var3", new Interval(Long.MIN_VALUE, -2L));
-    intervals = intervals.putAndCopy("fun::var1", new Interval(0L, 10L));
-    intervals = intervals.putAndCopy("fun::varB", new Interval(8L, Long.MAX_VALUE));
-    intervals = intervals.putAndCopy("fun::varC", new Interval(-15L, -3L));
+    intervals = intervals.putAndCopy("var1", new Interval(Long.MIN_VALUE, (long) 5));
+    intervals = intervals.putAndCopy("var2", new Interval((long) -7, Long.MAX_VALUE));
+    intervals = intervals.putAndCopy("var3", new Interval(Long.MIN_VALUE, (long) -2));
+    intervals = intervals.putAndCopy("fun::var1", new Interval((long) 0, (long) 10));
+    intervals = intervals.putAndCopy("fun::varB", new Interval((long) 8, Long.MAX_VALUE));
+    intervals = intervals.putAndCopy("fun::varC", new Interval((long) -15, (long) -3));
 
     IntervalAnalysisState iStateTest = new IntervalAnalysisState(intervals, referenceMap);
     IntervalRequirementsTranslator iReqTransTest =
@@ -241,7 +241,7 @@ public class TranslatorTest {
     // Test method convertToFormula() with another IntervalAnalysisState
     intervals = PathCopyingPersistentTreeMap.of();
     referenceMap = PathCopyingPersistentTreeMap.of();
-    intervals = intervals.putAndCopy("var1", new Interval(0L, Long.MAX_VALUE));
+    intervals = intervals.putAndCopy("var1", new Interval((long) 0, Long.MAX_VALUE));
     IntervalAnalysisState anotherIStateTest = new IntervalAnalysisState(intervals, referenceMap);
 
     convertedToFormula = iReqTransTest.convertToFormula(anotherIStateTest, ssaTest, null);
